+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Change Log
 
+## [0.0.3] - 23-03-2024
+
+### Added
+
+- Completed register page
+
+### Changed
+
+- Optimized the validation logic in the register page
+
+## [0.0.2] - 21-03-2024
+
+### Added
+
+- Added a shadcn toast to handle errors
+
+### Changed
+
+- Modified the create user controller
+- Switched from the hard coded toast
+
+## [0.0.1] - 17-03-2024
+
+### Added
+
+- Added an api function call for registering a user
+- Added a create user controller
+- Installed dotenv
+- Setup services folder for api endpoints
+- Added ShadCn ui
+- Added both error and value states to the form
+- Added an alert for displaying errors
+- Added a validation function
+
+### Changed
+
+- Changed naming conventions for services and controllers
+- Modified the form submission function
+-
+
 ## [0.0.0] - 16-03-2024
 
 ### Added